requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h516909a_2
ca-certificates=2020.6.24=0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38_0
cudatoolkit=10.2.89=hfd86e86_1
dbus=1.13.6=he372182_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.2=h5ab3b9f_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.65.0=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.13=h79a8f9a_0
graphite2=1.3.13=he1b5a44_1001
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_101
icu=64.2=he1b5a44_1
intel-openmp=2020.1=217
jasper=1.900.1=h07fcdf6_1006
jpeg=9d=h516909a_0
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=hdf63c60_15
libiconv=1.16=h516909a_0
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm9=9.0.1=he513fc3_1
libopencv=4.4.0=py38_2
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he6710b0_1
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
ninja=1.10.0=py38hfd86e86_0
nspr=4.27=he1b5a44_1
nss=3.55=he751ad9_0
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
olefile=0.46=py_0
opencv=4.4.0=py38_2
openh264=2.1.1=h8b12597_0
openssl=1.1.1g=h7b6447c_0
pcre=8.44=he1b5a44_0
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.2=py38_0
pixman=0.38.0=h516909a_1003
pthread-stubs=0.4=h14c3975_1001
py-opencv=4.4.0=py38h23f93f0_2
python=3.8.3=cpython_he5300dc_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.2.89_cudnn7.6.5_0
qt=5.12.5=hd8c4c69_1
readline=8.0=h7b6447c_0
scipy=1.5.0=py38h0b6359f_0
setuptools=49.6.0=py38_0
six=1.15.0=py_0
sqlite=3.32.3=h62c20be_0
tk=8.6.10=hbc83047_0
torchvision=0.7.0=py38_cu102
wheel=0.34.2=py38_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.11=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
